Rename controller package
